backend-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
emails==0.6
pydantic[email]==2.5.2
aiofiles==23.2.1
jinja2==3.1.2
bcrypt==4.0.1
requests==2.31.0
urllib3==1.26.20
chardet==5.2.0
charset-normalizer==3.3.2
